neurova_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
psycopg2-binary>=2.9
django-storages>=1.14
boto3>=1.34
reportlab>=4.2
python-dotenv>=1.0
